Create index.html
First design. The color scheme: yellow and baby blue
Web Page is dynamic, always centered on mobile and desktop. background, I want a soft ball field it looks like a big diamond  with 4 squares on each end of the diamond . there is a small yellow circle that move randomly  around the diamond when you click the screen . put the word softball in bold pink and green colors .